requirements.txt
django==1.11.1
django-jet==1.0.6
django-phonenumber-field==1.3.0
Pillow=4.1.1
